refactor: move validator to shared helpers
